Move ImplementationSnapshot to :snapshots
