+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/ImplementationAwareTaskAction.java
@@ -19,7 +19,7 @@ package org.gradle.api.internal.tasks;
 import org.gradle.api.Action;
 import org.gradle.api.Task;
 import org.gradle.internal.hash.ClassLoaderHierarchyHasher;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 
 public interface ImplementationAwareTaskAction extends Action<Task> {
