+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/SnapshotTaskInputsBuildOperationResultTest.groovy
@@ -23,7 +23,7 @@ import org.gradle.internal.execution.caching.CachingInputs
 import org.gradle.internal.execution.caching.CachingState
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint
 import org.gradle.internal.hash.HashCode
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot
+import org.gradle.internal.snapshot.ImplementationSnapshot
 import spock.lang.Specification
 
 class SnapshotTaskInputsBuildOperationResultTest extends Specification {
