+++ subprojects/execution/src/main/java/org/gradle/internal/execution/UnitOfWork.java
@@ -28,7 +28,7 @@ import org.gradle.internal.fingerprint.FileCollectionFingerprint;
 import org.gradle.internal.fingerprint.overlap.OverlappingOutputs;
 import org.gradle.internal.reflect.TypeValidationContext;
 import org.gradle.internal.snapshot.FileSystemSnapshot;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 
 import javax.annotation.Nullable;
 import java.io.File;
