+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/ExecutionState.java
@@ -19,8 +19,8 @@ package org.gradle.internal.execution.history;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 import org.gradle.internal.snapshot.ValueSnapshot;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
 
 public interface ExecutionState {
     /**
