+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/changes/ImplementationChanges.java
@@ -18,7 +18,7 @@ package org.gradle.internal.execution.history.changes;
 
 import com.google.common.collect.ImmutableList;
 import org.gradle.api.Describable;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 
 import javax.annotation.Nullable;
