+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/impl/AbstractExecutionState.java
@@ -20,8 +20,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.internal.execution.history.ExecutionState;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 import org.gradle.internal.snapshot.ValueSnapshot;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
 
 public class AbstractExecutionState<C extends FileCollectionFingerprint> implements ExecutionState {
     protected final ImplementationSnapshot implementation;
