+++ subprojects/execution/src/testFixtures/groovy/org/gradle/internal/execution/TestExecutionHistoryStore.java
@@ -25,8 +25,8 @@ import org.gradle.internal.execution.history.impl.DefaultAfterPreviousExecutionS
 import org.gradle.internal.execution.history.impl.SerializableFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 import org.gradle.internal.snapshot.ValueSnapshot;
-import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
 
 import java.util.HashMap;
 import java.util.Map;
