+++ subprojects/model-core/src/main/java/org/gradle/internal/snapshot/impl/AttributeDefinitionSnapshot.java
@@ -19,6 +19,7 @@ package org.gradle.internal.snapshot.impl;
 import org.gradle.api.attributes.Attribute;
 import org.gradle.internal.hash.ClassLoaderHierarchyHasher;
 import org.gradle.internal.hash.Hasher;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 
 public class AttributeDefinitionSnapshot extends AbstractIsolatableScalarValue<Attribute<?>> {
