+++ subprojects/model-core/src/main/java/org/gradle/internal/snapshot/impl/SnapshotSerializer.java
@@ -22,6 +22,7 @@ import org.gradle.internal.serialize.Decoder;
 import org.gradle.internal.serialize.Encoder;
 import org.gradle.internal.serialize.HashCodeSerializer;
 import org.gradle.internal.serialize.Serializer;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
 import org.gradle.internal.snapshot.ValueSnapshot;
 
 public class SnapshotSerializer extends AbstractSerializer<ValueSnapshot> {
@@ -46,7 +47,7 @@ public class SnapshotSerializer extends AbstractSerializer<ValueSnapshot> {
     private static final int DEFAULT_SNAPSHOT = 18;
 
     private final HashCodeSerializer serializer = new HashCodeSerializer();
-    private final Serializer<ImplementationSnapshot> implementationSnapshotSerializer = new ImplementationSnapshot.SerializerImpl();
+    private final Serializer<ImplementationSnapshot> implementationSnapshotSerializer = new ImplementationSnapshotSerializer();
 
     @Override
     public ValueSnapshot read(Decoder decoder) throws Exception {
