+++ subprojects/snapshots/src/main/java/org/gradle/internal/snapshot/impl/LambdaImplementationSnapshot.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.snapshot.impl;
+
+import org.gradle.internal.hash.HashCode;
+import org.gradle.internal.hash.Hasher;
+import org.gradle.internal.snapshot.ImplementationSnapshot;
+
+import javax.annotation.Nullable;
+
+public class LambdaImplementationSnapshot extends ImplementationSnapshot {
+
+    public LambdaImplementationSnapshot(String typeName) {
+        super(typeName);
+    }
+
+    @Override
+    public void appendToHasher(Hasher hasher) {
+        hasher.markAsInvalid(getUnknownReason());
+    }
+
+    @Override
+    protected boolean isSameSnapshot(@Nullable Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        LambdaImplementationSnapshot that = (LambdaImplementationSnapshot) o;
+
+        return getTypeName().equals(that.getTypeName());
+    }
+
+    @Override
+    public HashCode getClassLoaderHash() {
+        return null;
+    }
+
+    @Override
+    public boolean isUnknown() {
+        return true;
+    }
+
+    @Override
+    @Nullable
+    public String getUnknownReason() {
+        return "was implemented by the Java lambda '" + getTypeName() + "'. Using Java lambdas is not supported, use an (anonymous) inner class instead.";
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        return getTypeName().hashCode();
+    }
+
+    @Override
+    public String toString() {
+        return getTypeName();
+    }
+}
